Move "packs" to "kit"
